requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
asgiref==3.5.2
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
DateTime==4.7
Deprecated==1.2.13
discord-webhook==0.17.0
discord.py==1.7.3
Django==4.1.1
dnspython==2.2.1
exceptiongroup==1.0.0rc9
future==0.18.2
gTTS==2.2.4
h11==0.14.0
idna==3.3
image==1.5.33
multidict==6.0.2
numpy==1.23.3
outcome==1.2.0
packaging==21.3
pefile==2021.9.3
Pillow==9.2.0
pyarmor==7.3.6
pycparser==2.21
pycryptodome==3.15.0
pyfiglet==0.8.post1
pygame==2.1.2
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.1
pylock==0.3
pymongo==4.2.0
pyparsing==3.0.7
pyperclip==1.8.2
pypresence==4.2.1
PySocks==1.7.1
tqdm==4.64.1
tzdata==2022.4
urllib3==1.26.12
wrapt==1.13.3
wsproto==1.2.0
yarl==1.7.2
zope.interface==5.4.0
pytz==2022.2.1
pywin32-ctypes==0.2.0
redis==4.1.3
requests==2.28.1
sqlparse==0.4.3